pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seeed-mr60bha2"
version = "1.0.0"
description = "Python library for Seeed MR60BHA2 mmWave heart rate and breath monitoring sensor"
readme = "README.md"
authors = [
    {name = "AB-Engineering"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
]
keywords = ["seeed", "mmwave", "sensor", "heart-rate", "breath-rate", "raspberry-pi", "iot"]
requires-python = ">=3.7"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.9",
]

[project.urls]
Homepage = "https://github.com/AB-Engineering/MR60BHXX-python-driver"
Documentation = "https://github.com/AB-Engineering/MR60BHXX-python-driver"
Repository = "https://github.com/AB-Engineering/MR60BHXX-python-driver"
Issues = "https://github.com/AB-Engineering/MR60BHXX-python-driver/issues"

[tool.setuptools]
packages = ["seeed_mr60bha2"]

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
